requirements.txt

beautifulsoup4==4.12.2
fqdn==1.5.1
ftfy==6.2.0
Jinja2==3.1.2
jmespath==1.0.1
json5==0.9.14
nltk==3.2.4
numpy==1.23.5
onnxruntime==1.19.2
opencv-python==4.7.0.72
opencv-python-headless==4.7.0.72
Pillow==9.5.0
python-json-logger==2.0.7
python-slugify==8.0.1
pytorch-warmup==0.1.1
pytorchvideo==0.1.5
pytz==2023.3
PyYAML==5.4.1
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
torch==2.0.1+cu118
torchaudio==2.0.2+cu118
torchdata==0.6.1
torchtext==0.15.2
torchvision==0.15.2+cu118
transformers==4.44.2
zipp==3.15.0
